Rename `html_email.py` to `renderers.py`
It’s going to be used for more than just HTML emails.
